Authorizing Calls to APIs with OAuth2: Filtering Data in the API Based on Caller Identity

+++ openid-and-oauth2-securing-angular-apps/SecuringAngularApps.API/SecuringAngularApps.API/Controllers/ProjectsController.cs
@@ -37,12 +37,14 @@ namespace SecuringAngularApps.API.Controllers
 
         // GET: api/Projects/5
         [HttpGet("{id}")]
-        public IActionResult GetProject([FromRoute] int id)
+        public async Task<IActionResult> GetProject([FromRoute] int id)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
+
+            if (!await ProjectEditAccessCheck(id, false)) return Forbid();
 
             var project = _context.Projects
                 .Include("UserPermissions")
@@ -82,6 +84,8 @@ namespace SecuringAngularApps.API.Controllers
             {
                 return BadRequest();
             }
+
+            if (!await ProjectEditAccessCheck(id, true)) return Forbid();
 
             _context.Entry(project).State = EntityState.Modified;
 
@@ -147,6 +151,8 @@ namespace SecuringAngularApps.API.Controllers
         {
             var existing = await _context.Milestones.FirstOrDefaultAsync(m => m.Id == milestone.Id);
             if (existing != null) return StatusCode(409);
+            if (!await MilestoneAccessCheck(existing)) return Forbid();
+            milestone.MilestoneStatusId = 1;
             _context.Milestones.Add(milestone);
             await _context.SaveChangesAsync();
             return CreatedAtAction("GetProject", new { id = milestone.ProjectId }, milestone);
@@ -157,6 +163,7 @@ namespace SecuringAngularApps.API.Controllers
         {
             var item = await _context.Milestones.FirstOrDefaultAsync(m => m.Id == id);
             if (item == null) return NotFound();
+            if (!await MilestoneAccessCheck(item)) return Forbid();
             _context.Milestones.Remove(item);
             await _context.SaveChangesAsync();
             return Ok();
@@ -168,10 +175,27 @@ namespace SecuringAngularApps.API.Controllers
             if (milestone.Id != id) return BadRequest();
             var item = await _context.Milestones.FirstOrDefaultAsync(ms => ms.Id == id);
             if (item == null) return NotFound();
+            if (!await MilestoneAccessCheck(item)) return Forbid();
             item.MilestoneStatusId = milestone.MilestoneStatusId;
             item.Name = milestone.Name;
             await _context.SaveChangesAsync();
             return Ok(milestone);
+        }
+
+        private async Task<bool> MilestoneAccessCheck(Milestone item)
+        {
+            var userId = this.User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var perm = await _context.UserPermissions.FirstOrDefaultAsync(up =>
+                up.ProjectId == item.ProjectId && up.UserProfileId == userId);
+            return (perm != null && perm.Value == "Edit");
+        }
+
+        private async Task<bool> ProjectEditAccessCheck(int projectId, bool edit)
+        {
+            var userId = this.User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var userAccess = await _context.UserPermissions.FirstOrDefaultAsync(up =>
+                up.ProjectId == projectId && up.UserProfileId == userId);
+            return (userAccess != null && (edit ? userAccess.Value == "Edit" : true));
         }
 
         [HttpGet("MilestoneStatuses")]
